requirements.txt
requests
bs4
datetime
pandas
psycopg2-binary
python-dotenv
sqlalchemy